Implements modification of predefined styles.
For compatibility with current API and style definitions (as inline CSS in styles.csv), I chose to match and replace the relevant bits with regular expressions. This might not be the most error save approach, but keeps code changes smaller and allows for backwards compatibility.

drop_name and drop_html pass the relevant arguments to get_css_styles.